Fixed a bug that was causing the sensors to throw an exception.
In the mqtt node in this flow, I missed a character in the address. This
caused and error to be thrown by the gateway that made it so the gateway
would not respond to the requests. This bug was found by Rainer, a WSN
employee. Make sure the address are correct (24 characters long).