Change the name and complete working sensor
I changed the name of the file to identify it as solely for that
specific sensor. The flow is fully functioning. It can gather periodic
data from the sensor and preprocess it for out of bounds temperature. I
can periodically grab all the manual LED changes and act on them
accordingly.