Adding Maunal LED concurrency solution and second senor flow
I added in a queue with a trigger release to deal with multiple manual
LED change commands. Each command grabed from the iotmms is stored in a
queue and they are released one at a time when there is a trigger from
one of the commands getting to the end of the flow.

I also copied and modifed the exsg flow to be used for the exom sensor
so it has a little testing left, but it should work just like the exsg
flow.